fix(tests): address code review for --models-dir (#8362)
- Move HF offline helpers to tests/hf_cache.py so tests can import them
  without fragile from-conftest imports
- Set HF_HUB_OFFLINE=1 directly in _apply_models_dir_env; Mistral patch
  call is now annotated as sitecustomize-only side-effect
- Remove @pytest.mark.gpu_0 from all unit tests in test_models_dir_flag.py
- Remove redundant test_apply_does_not_set_transformers_cache (covered by
  test_apply_bare_cache_layout)
- Use tmp_path-derived path in nonexistent-dir subprocess test
- Add test_restore_handles_missing_tmp_cache for OSError branch
- Trim --models-dir help text to two lines; full docs in pytest-guidelines.md
- Remove DYNAMO_MODELS_DIR env-var escape hatch from predownload fixtures;
  require --models-dir to activate offline mode
